Extract `system.system` model to its own file
This model is intended to be the same between both the Data Source and
the Resource. It shouldn't live in the Data Source file. So, we put it
in its own file.
